+++ CHANGES.rst
@@ -46,6 +46,9 @@ Changelog
 -   Use nodes dedicated to online PE also for playground events so that the
     test runs do not get stuck due to the lack of resources.
 
+-   Use ``EM_Selected`` to freeze the preferred event of a superevent and launch
+    a preliminary alert.
+
 0.8.7 (2019-09-14)
 ------------------